+++ src/Exception/TemplateRenderExceptionInterface.php
@@ -3,23 +3,23 @@
 namespace Dhii\Output\Exception;
 
 use Dhii\Output\ContextAwareInterface;
-use Dhii\Output\ContextRendererInterface;
+use Dhii\Output\TemplateInterface;
 
 /**
  * An exception that occurs when cannot render with a context.
  *
- * @since 0.1
+ * @since 0.2
  */
-interface ContextRenderExceptionInterface extends
+interface TemplateRenderExceptionInterface extends
     CouldNotRenderExceptionInterface,
     ContextAwareInterface
 {
     /**
      * {@inheritdoc}
      *
-     * @since 0.1
+     * @since 0.2
      * 
-     * @return ContextRendererInterface The renderer.
+     * @return TemplateInterface The renderer.
      */
     public function getRenderer();
 }