+++ src/RendererInterface.php
@@ -2,6 +2,7 @@
 
 namespace Dhii\Output;
 
+use Dhii\Util\String\StringableInterface as Stringable;
 use Dhii\Output\Exception\CouldNotRenderExceptionInterface;
 use Dhii\Output\Exception\RendererExceptionInterface;