+++ src/TemplateInterface.php
@@ -3,7 +3,7 @@
 namespace Dhii\Output;
 
 use Dhii\Util\String\StringableInterface as Stringable;
-use Dhii\Output\Exception\ContextRenderExceptionInterface;
+use Dhii\Output\Exception\TemplateRenderExceptionInterface;
 use Dhii\Output\Exception\RendererExceptionInterface;
 
 /**
@@ -13,20 +13,20 @@ use Dhii\Output\Exception\RendererExceptionInterface;
  * all the data necessary for rendering at the time of
  * rendering.
  *
- * @since 0.1
+ * @since 0.2
  */
-interface ContextRendererInterface extends RendererInterface
+interface TemplateInterface extends RendererInterface
 {
     /**
      * Produce output based on context.
      *
-     * @since 0.1
+     * @since 0.2
      *
      * @param mixed|null $context The context;
      *                            something that can provide more information on how to perform rendering.
      *
-     * @throws ContextRenderExceptionInterface If cannot render.
-     * @throws RendererExceptionInterface      Any other problem related to the renderer.
+     * @throws TemplateRenderExceptionInterface If cannot render.
+     * @throws RendererExceptionInterface       Any other problem related to the renderer.
      *
      * @return string|Stringable The output.
      */