+++ test/unit/TemplateInterfaceTest.php
@@ -2,27 +2,27 @@
 
 namespace Dhii\Output\UnitTest;
 
-use Dhii\Output\ContextRendererInterface as TestSubject;
+use Dhii\Output\TemplateInterface as TestSubject;
 use Xpmock\TestCase;
 
 /**
  * Tests {@see TestSubject}.
  *
- * @since 0.1
+ * @since 0.2
  */
-class ContextRendererInterfaceTest extends TestCase
+class TemplateInterfaceTest extends TestCase
 {
     /**
      * The class name of the test subject.
      *
-     * @since 0.1
+     * @since 0.2
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Output\ContextRendererInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Output\TemplateInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
-     * @since 0.1
+     * @since 0.2
      *
      * @return TestSubject A new instance of the test subject.
      */
@@ -38,7 +38,7 @@ class ContextRendererInterfaceTest extends TestCase
     /**
      * Tests whether a valid instance of the test subject can be created.
      *
-     * @since 0.1
+     * @since 0.2
      */
     public function testCanBeCreated()
     {